refactor datastore
break out interfaces for plot and tree data access.
add plot error data access methods
