+++ source/FScruiser.Core/Models/Abstractions/IHasTreeID.cs
