+++ source/FScruiser.Core/Models/PlotError.cs
@@ -1,0 +1,15 @@
+﻿using FMSC.ORM.EntityModel.Attributes;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FScruiser.Models
+{
+    public class PlotError : Error_Base
+    {
+        [Field("PlotID")]
+        public string PlotID { get; set; }
+    }
+}
