+++ source/FScruiser.Core/Models/TreeError.cs
@@ -1,30 +1,11 @@
 ﻿using FMSC.ORM.EntityModel.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using FScruiser.Models;
 
 namespace FScruiser.Models
 {
-    public class TreeError : IHasTreeID
+    public class TreeError : Error_Base, IHasTreeID
     {
         [Field("TreeID")]
         public string TreeID { get; set; }
-
-        [Field("Level")]
-        public string Level { get; set; }
-
-        [Field("Message")]
-        public string Message { get; set; }
-
-        [Field("Field")]
-        public string Field { get; set; }
-
-        [Field("Resolution")]
-        public string Resolution { get; set; }
-
-        [Field("ResolutionInitials")]
-        public string ResolutionInitials { get; set; }
     }
 }
