+++ source/FScruiser.Core/Models/TreeMeasurment.cs
@@ -1,10 +1,5 @@
 ﻿using FMSC.ORM.EntityModel.Attributes;
 using FScruiser.Util;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FScruiser.Models
 {
