+++ source/FScruiser.Core/Models/TreeStub.cs
@@ -1,9 +1,5 @@
 ﻿using FMSC.ORM.EntityModel.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using FScruiser.Models;
 
 namespace FScruiser.Models
 {
