+++ source/FScruiser.Core/Services/Abstractions/IDialogService.cs
