+++ source/FScruiser.Core/Services/Abstractions/ITallySettingsDataService.cs
